Implemented seeding organisations (#30)
* Implemented seeding organisations

* Implemented getting tokens by external clients

* Fixed Codacy remark

* Started implementing adding products

* Adjusted model builders

* Implemented catalog view models

* Fixed double quotes

* Implemented storefront catalog and product labels

* Updated libraries

* Implemented docker compose for elasticsearch

* Installed ElasticSearch

* Implemented getting products from elastic search

* Implemented integration with elasticsearch

* Fixed pagination of category products

* Fixed category catalog

* Implemented getting paged category products

* Implemented creating new media file

* Added migrations for media

* Added catalog migrations

* Updated Catalog context

* Added Identity migrations

* Removed product's SKU a mandatory field

* Implemented creating new products in catalog

* Removed redundant usings

* Implemented uploading files

* Implemented checksum checking

* Implemented calculating checksum for seeded media

* Fixed prodcut request model

* Added primary product id mapping

* Fixed adding images, files and videos via api

* Implemented adding products from API

* Removed request size limit

* Added protected and is new properties to product and removed is default from language

* Add migrations for IsProtected

* Fixed importing documents

* Removed redunant using

* Implemented creating and indexing products

* Implemented listing products in categories

* Implemented displaying products on category pages

* Updated nuget packages

* Implemented scaling images

* Fixed SSR for Products

* Fixed pagination

* Implemented product detail page

* Implemented product variants

* Updated categories list and hero slider items

* Fixed product detail page

* Implemented pagination for category pages

* Started implementing brand detail page

* Implemented empty brand detail page

* Implemented brand catalog

* Fixed code smells and issues

* Implemented getting logo and favicon from media api

* Fixed codacy remarks

* Fixed possible null reference exception

* Reduced methods complexity

* Fixed brand constants

* Reduced methods complexity

* Implemented Swagger for Identity API

* Implemented identity migrations

* Implemented brand repository and seller controller

* Implemented seller check

* Added DI for SellerService

* Updated .NET Core to the latest version

* Fixed Logo frontend

* Fixed getting product variants

* Implemented cookie consent, privacy policy and regulations

* Finished off the cookie consent

* Implemented Codacy remarks

* Fixed styling for Privacy Policy and Regulations pages

* Fixed logo url

* Fixed header props

* Fixed regulations and privacy policy links

* Implemented files model builder

* Finished off downloading products and copying links

* Fixed products list for brand page

* Added getting files from brand

* Fixed possible null reference exception

* Removed redundant http accessor

+++ be/src/Infrastructure/GenericRepository/Foundation.GenericRepository/Paginations/Pagination.cs
@@ -4,7 +4,7 @@ namespace Foundation.GenericRepository.Paginations
 {
     public class Pagination
     {
-        public int TotalItems { get; }
+        public long TotalItems { get; }
 
         public int PageSize { get; }
